analysis: src: make_analysis: expand lists of interest to find replace for linux-sound
In a meeting, we found out that `linux-sound` may not be the best
example of a list to be included in the plot for the analysis. We
determined that it would be better to choose a big pool of lists to
first assess interesting trends then choose. Expand the set of lists in
the plot for this purpose.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>